Refactor: Split bloated app.js into ES Modules (api.js, ui.js, app.js)
