+++ templates/index.html
@@ -59,6 +59,6 @@
         </div>
     </div>
 
-    <script src="/static/js/app.js"></script>
+    <script type="module" src="/static/js/app.js"></script>
 </body>
 </html>
